Merge remote-tracking branch 'origin/3.4.x'

+++ wars/api-pagamento/src/main/java/it/govpay/pagamento/v1/beans/ListaPendenze.java
@@ -7,7 +7,7 @@ import it.govpay.core.beans.Lista;
 
 public class ListaPendenze extends Lista<Pendenza> {
 	
-	public ListaPendenze(List<Pendenza> pagamentiPortale, URI requestUri, long count, long pagina, long limit) {
+	public ListaPendenze(List<Pendenza> pagamentiPortale, URI requestUri, Long count, Integer pagina, Integer limit) {
 		super(pagamentiPortale, requestUri, count, pagina, limit);
 	}
