+++ wars/api-pagamento/src/main/java/it/govpay/pagamento/v2/beans/ListaDomini.java
@@ -7,7 +7,7 @@ import it.govpay.core.beans.Lista;
 
 public class ListaDomini extends Lista<Dominio> {
 	
-	public ListaDomini(List<Dominio> domini, URI requestUri, long count, long pagina, long limit) {
+	public ListaDomini(List<Dominio> domini, URI requestUri, Long count, Integer pagina, Integer limit) {
 		super(domini, requestUri, count, pagina, limit);
 	}
