+++ wars/api-pagamento/src/main/java/it/govpay/pagamento/v2/beans/ListaTipiPendenza.java
@@ -7,7 +7,7 @@ import it.govpay.core.beans.Lista;
 
 public class ListaTipiPendenza extends Lista<TipoPendenza> {
 	
-	public ListaTipiPendenza(List<TipoPendenza> tipiEntrata, URI requestUri, long count, long pagina, long limit) {
+	public ListaTipiPendenza(List<TipoPendenza> tipiEntrata, URI requestUri, Long count, Integer pagina, Integer limit) {
 		super(tipiEntrata, requestUri, count, pagina, limit);
 	}
