+++ wars/api-ragioneria/src/main/java/it/govpay/ragioneria/v1/beans/ListaFlussiRendicontazione.java
@@ -11,7 +11,7 @@ public class ListaFlussiRendicontazione extends Lista<FlussoRendicontazioneIndex
 		super();
 	}
 	
-	public ListaFlussiRendicontazione(List<FlussoRendicontazioneIndex> flussiRendicontazione, URI requestUri, long count, long pagina, long limit) {
+	public ListaFlussiRendicontazione(List<FlussoRendicontazioneIndex> flussiRendicontazione, URI requestUri, Long count, Integer pagina, Integer limit) {
 		super(flussiRendicontazione, requestUri, count, pagina, limit);
 	}
