+++ wars/api-ragioneria/src/main/java/it/govpay/ragioneria/v2/beans/Riconciliazioni.java
@@ -7,7 +7,7 @@ import it.govpay.core.beans.Lista;
 
 public class Riconciliazioni extends Lista<RiconciliazioneIndex> {
 	
-	public Riconciliazioni(List<RiconciliazioneIndex> incassi, URI requestUri, long count, long offset, long limit) {
+	public Riconciliazioni(List<RiconciliazioneIndex> incassi, URI requestUri, Long count, Integer offset, Integer limit) {
 		super(incassi, requestUri, count, offset, limit);
 	}
